perf(core-web-vitals): optimize fonts and caching
