+++ next.config.js
@@ -1,13 +1,33 @@
+const withBundleAnalyzer = require('@next/bundle-analyzer')({
+  enabled: process.env.ANALYZE === 'true',
+});
+
 const redirects = async () => ([
   { source: '/post', destination: '/', permanent: false },
   { source: '/find', destination: '/', permanent: false }, // TEMP until cache settles
   { source: '/jobs', destination: '/', permanent: false }, // just in case
 ]);
 
-/** @type {import('next').NextConfig} */
-const nextConfig = {
+module.exports = withBundleAnalyzer({
   reactStrictMode: true,
+  images: { formats: ['image/avif', 'image/webp'] },
   eslint: { ignoreDuringBuilds: true }, // prevent CI failing on eslint/parser fetch
+  async headers() {
+    return [
+      {
+        source: '/(.*)\\.(?:js|css|woff2|png|jpg|jpeg|gif|svg|webp|avif)$',
+        headers: [
+          { key: 'Cache-Control', value: 'public, max-age=31536000, immutable' },
+        ],
+      },
+      {
+        source: '/favicon-:size.png',
+        headers: [
+          { key: 'Cache-Control', value: 'public, max-age=31536000, immutable' },
+        ],
+      },
+    ];
+  },
   async redirects() {
     const base = await redirects();
     return [
@@ -16,6 +36,4 @@ const nextConfig = {
       { source: '/signup', destination: '/', permanent: true },
     ];
   },
-};
-
-module.exports = nextConfig;
+});
