+++ package.json
@@ -15,6 +15,8 @@
     "qa:report": "playwright show-report",
     "build:prod": "next build",
     "start:prod": "next start -p 3000",
+    "analyze": "ANALYZE=true next build",
+    "build:profile": "NODE_OPTIONS='--trace-deopt --trace_gc' next build",
     "seed": "node scripts/seed.mjs",
     "postinstall": "node -e \"process.env.NODE_ENV==='production'?process.exit(0):0\" || npx playwright install --with-deps || true"
   },
@@ -47,7 +49,8 @@
     "fast-glob": "^3.3.2",
     "start-server-and-test": "^2.0.0",
     "@supabase/supabase-js": "^2.45.0",
-    "tsx": "^4.7.0"
+    "tsx": "^4.7.0",
+    "@next/bundle-analyzer": "^15.0.0"
   },
   "engines": {
     "node": ">=18.17"
