+++ pages/index.tsx
@@ -44,3 +44,7 @@ export default function Home() {
     </Card>
   );
 }
+
+export async function getStaticProps() {
+  return { props: {}, revalidate: 60 };
+}
